imported_pip_requirements.txt
beautifulsoup4==4.6.0
Django==1.10
django-app-namespace-template-loader==0.4.1
django-blog-zinnia==0.18.1
django-ckeditor==5.3.0
django-contrib-comments==1.8.0
django-disqus==0.5
django-js-asset==0.1.1
django-mptt==0.8.7
django-photologue==3.7
django-sortedm2m==1.5.0
django-tagging==0.4.5
django-taggit==0.22.1
-e git://github.com/Fantomas42/django-xmlrpc.git@6cf59c555b207de7ecec75ac962751e8245cf8c9#egg=django_xmlrpc
docutils==0.14
ExifRead==2.1.2
gunicorn==19.7.1
mock==2.0.0
mots-vides==2015.5.11
mysqlclient==1.3.12
olefile==0.44
pbr==3.1.1
Pillow==4.3.0
pyparsing==2.2.0
pytz==2017.2
regex==2017.9.23
six==1.11.0
